+++ Tests/NukeExtensionsTests/ImageViewExtensionsProgressiveDecodingTests.swift
@@ -36,7 +36,7 @@ import Foundation
             $0.isProgressiveDecodingEnabled = true
             $0.progressiveDecodingInterval = 0
             $0.isStoringPreviewsInMemoryCache = true
-            $0.imageProcessingQueue.maxConcurrentOperationCount = 1
+            $0.imageProcessingQueue.maxConcurrentTaskCount = 1
         }
     }